fix: fix regex for filtering credentials from logs
The ending "=" character of a base64 encoded string is padding. When the
input length is a multiple of three (which is the case for our account
id and license key), the output would not have the padding character.
This results in our credentials not being filtered in the logs.

This fixes the regex by removing the incorrect assumption.

https://en.wikipedia.org/wiki/Base64#Output_padding